Calculation of actual evaporation and GPP using the PML_V2 for global land surface for each time step (16-day).

This algorithm fits to high resolution data data in GEE.

PML(
  data,
  par = NULL,
  CO2 = NULL,
  IGBPcode = NULL,
  ratio = 1,
  ...,
  scale = c("site", "regional"),
  V2 = TRUE,
  return.Es = TRUE
)

Arguments

data

A data.frame with the columns of

Prcp : Precipitation, mm

Tavg : daily average temperature, deg C

Rs : shortwave radiation, W m-2

Pa : surface air pressure, kPa

U2 : wind speed, m/s

LAI : m2 m-2

(can be calculated by other variables)

Rn : net solar radiation, W m-2

albedo

emiss

Rln, W m-2 (plus Rs, Tavg): Rn = get_Rn(Rs, Rln, Tavg, albedo, emiss)

VPD : Vapor pressure deficit, kPa

q, kg/kg (plus Pa): VPD = get_VPD(Pa, q)

(optional)

CO2 : about 380-570, umol mol-1

IGBPname: If not specified parameters, this column is necessary for matching parameter.

(optional, for speeding up optimization)

lambda :

rho_a : air density, [g m-3]

gama :

epsilon :

par

PML_V2 parameters (see Zhang 2019 for details).

hc : canopy height (m)

Alpha : initial photochemical efficiency, 0.02-0.08

m : the initla slope of CO2 response curve [umol m-2 s-1]/[umol mol-1]

Am_25 : 10 - 150

D0 : the value of VPD when stomtal conductance is reduced

kQ : 0.7,

kA : 0.7, light extinction coefficient

S_sls :

fER0 :

VPDmin :

VPDmax :

CO2

A constant number. If provided, CO2 in the data will be replaced.

IGBPcode

IGBP006 code

ratio

land cover ratio

...

ignored

scale

one of c("site", "regional").

V2

Boolean, default TRUE. If TRUE PML_V2 used, otherwise PML_V1.

return.Es

Because fval_soil need moving average in temporal. We can't get it directly if the data is Spatial pixels.

Value

GPP: gC m-2 d-1, Gross primary product

Ec : mm d-1, Vegetation transpiration

Es : mm d-1, Soil evaporation |

Ei : mm d-1, Interception from vegetation canopy

Es_eq: mm d-1, Soil evaporation at equilibrium

Pi: mm d-1, Prcp - Ei

References

Zhang, Y., Kong, D., Gan, R., Chiew, F. H. S., McVicar, T. R., Zhang, Q., & Yang, Y. (2019). Coupled estimation of 500 m and 8-day resolution global evapotranspiration and gross primary production in 2002–2017. Remote Sensing of Environment, 222, 165–182. doi:10.1016/j.rse.2018.12.031.

Kong, D., Zhang, Y., Gu, X., & Wang, D. (2019). A robust method for reconstructing global MODIS EVI time series on the Google Earth Engine. ISPRS Journal of Photogrammetry and Remote Sensing, 155, 13–24. https://doi.org/10.1016/j.isprsjprs.2019.06.014

basic units change, http://www.fao.org/docrep/X0490E/x0490e0i.htm